split llm module interfaces into individual files
