+++ convlab/base_models/llm/__init__.py
@@ -1,2 +1,5 @@
 from convlab.base_models.llm.base import LLM
-from convlab.base_models.llm.models import LLM_US, LLM_RG, LLM_DST, LLM_NLU, LLM_NLG
+from convlab.base_models.llm.user_simulator import LLM_US, LLM_RG
+from convlab.base_models.llm.nlu import LLM_NLU
+from convlab.base_models.llm.dst import LLM_DST
+from convlab.base_models.llm.nlg import LLM_NLG
